fix: now typedoc generates documentation
