+++ src/index.ts
@@ -1,32 +1,6 @@
-import { IOBuffer } from 'iobuffer';
-
-import { readDataBlock, Spectrum } from './dataBlock';
-import { LogBlock, readLogBlock } from './logBlock';
-import { Header, TheNewHeader, mainHeader } from './mainHeader';
-
-declare type InputData = ArrayBufferLike | ArrayBufferView | IOBuffer | Buffer;
-
-interface ParseResult {
-  meta: Header;
-  spectra: Spectrum[];
-  /* logs may be missing */
-  logs?: LogBlock;
-}
-
-/**
- * Parses an SPC file.
- *
- * @param buffer - SPC file buffer.
- * @return Object containing every information contained in the SPC file.
- */
-export function parse(buffer: InputData): ParseResult {
-  const ioBuffer = new IOBuffer(buffer);
-  const meta = mainHeader(ioBuffer);
-  const spectra = readDataBlock(ioBuffer, meta);
-  if (meta instanceof TheNewHeader) {
-    if (meta.logOffset && meta.logOffset !== 0) {
-      return { meta, spectra, logs: readLogBlock(ioBuffer, meta.logOffset) };
-    }
-  }
-  return { meta, spectra };
-}
+export * from './utility';
+export * from './types';
+export * from './mainHeader';
+export * from './logBlock';
+export * from './dataBlock';
+export * from './parse';
