+++ src/logBlock.ts
@@ -1,6 +1,6 @@
 import { IOBuffer } from 'iobuffer';
 
-interface MetaData {
+export interface MetaData {
   size: number;
   memorySize: number;
   textOffset: number;
@@ -17,9 +17,9 @@ export interface LogBlock {
 
 /**
  *
- * @param  buffer SPC buffer.
- * @param  logOffset Offset of the log (from mainHeader).
- * @return  Object containing log meta, data and text.
+ * @param buffer - SPC buffer.
+ * @param logOffset - Offset of the log (from mainHeader).
+ * @return Object containing log meta, data and text.
  */
 export function readLogBlock(buffer: IOBuffer, logOffset: number): LogBlock {
   let logHeader: MetaData = {
